Change return type of nodeMultiPass_ to void
The function wasn't returning a value, so LLVM put a ud2 instruction at
the end, so the function would crash instead of returning garbage.
Presumably this is a feature, intended to help you find undefined
behavior sooner. Didn't stop me from spending hours debugging, not
finding the problem, rebuilding the project with ubsan, and finally
fixing the problem by changing bool to void in two places.