[Gen Sched] Remove issue slot count, update SCW denominator.
Remove issue slot count in the machine model. The value should not be
set in the config file and it was not being used anyway. If it is
necessary to know this information the issue types can be queried. This
method is less error prone as we have discovered. Also change the spill
cost weight (SCW) to allow for lower values until the calculation can be
updated like we have planned. There is also a change to what constitutes
a simple machine model. The model must have no pipelined instructions
and a single issue type with only one slot.